+++ semantic-conventions/CHANGELOG.md
@@ -4,7 +4,8 @@ Please update the changelog as part of any significant pull request.
 
 ## Unreleased
 
-- Add changes here.
+- Allow customising the `jinja2.Environment` via command line arguments
+  ([#xx](https://github.com/open-telemetry/build-tools/pull/xx)).
 
 ## v0.12.0